Move LocalFileStorage to tests directory to fix webpack build issue.
