+++ tests/data-access/data-access-utils.ts
@@ -1,6 +1,6 @@
 import { IDatabaseContext } from "../../src/lib/common/data-access/database-context-interface";
 import { AppDatabaseContext } from "../../src/lib/data-access/app-database-context";
-import { LocalFileStorage } from "../../src/lib/data-access/local-file-storage";
+import { LocalFileStorage } from "./local-file-storage";
 import { SingleFileJsonDatabase } from "../../src/lib/data-access/single-file-json-database";
 
 export class DataAccessUtils {
